+++ src/main/java/com/ambegodas/maxpos/model/Product.java
@@ -13,14 +13,14 @@ import java.util.Date;
 public class Product {
 	
 	@Id @GeneratedValue
-	private int productId;
+	private long productId;
 	private String productName;
 	private String description;
 	private double price;
 	private double cost;
 	private String measuredIn ;
-	private long availableQty;
-	private long soldQty;
+	private int availableQty;
+	private int soldQty;
 	private Date addedDate;
 
 
@@ -35,7 +35,7 @@ public class Product {
 		this.price = price;
 	}
 
-	public int getProductId() {
+	public long getProductId() {
 		return productId;
 	}
 
@@ -83,19 +83,19 @@ public class Product {
 		this.measuredIn = measuredIn;
 	}
 
-	public long getAvailableQty() {
+	public int getAvailableQty() {
 		return availableQty;
 	}
 
-	public void setAvailableQty(long availableQty) {
+	public void setAvailableQty(int availableQty) {
 		this.availableQty = availableQty;
 	}
 
-	public long getSoldQty() {
+	public int getSoldQty() {
 		return soldQty;
 	}
 
-	public void setSoldQty(long soldQty) {
+	public void setSoldQty(int soldQty) {
 		this.soldQty = soldQty;
 	}
